Rename PersonContainsKeywordsPredicate to FilterPredicate
