+++ src/main/java/seedu/dengue/logic/commands/FindCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.dengue.commons.core.Messages;
 import seedu.dengue.model.Model;
-import seedu.dengue.model.predicate.PersonContainsKeywordsPredicate;
+import seedu.dengue.model.person.FilterPredicate;
 
 /**
  * Finds and lists all persons in Dengue Hotspot Tracker whose name contains any of the argument keywords.
@@ -20,9 +20,9 @@ public class FindCommand extends Command {
             + "Example: " + COMMAND_WORD + " alice bob charlie";
 
     // TODO: change to accept any predicate
-    private final PersonContainsKeywordsPredicate predicate;
+    private final FilterPredicate predicate;
 
-    public FindCommand(PersonContainsKeywordsPredicate predicate) {
+    public FindCommand(FilterPredicate predicate) {
         this.predicate = predicate;
     }
