+++ src/main/java/seedu/dengue/model/person/FilterPredicate.java
@@ -1,18 +1,17 @@
-package seedu.dengue.model.predicate;
+package seedu.dengue.model.person;
 
 import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.dengue.commons.util.StringUtil;
-import seedu.dengue.model.person.Person;
 
 /**
  * Tests that a {@code Person}'s {@code Name} matches any of the keywords given.
  */
-public class PersonContainsKeywordsPredicate implements Predicate<Person> {
+public class FilterPredicate implements Predicate<Person> {
     private final List<String> keywords;
 
-    public PersonContainsKeywordsPredicate(List<String> keywords) {
+    public FilterPredicate(List<String> keywords) {
         this.keywords = keywords;
     }
 
@@ -35,7 +34,7 @@ public class PersonContainsKeywordsPredicate implements Predicate<Person> {
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof PersonContainsKeywordsPredicate // instanceof handles nulls
-                && keywords.equals(((PersonContainsKeywordsPredicate) other).keywords)); // state check
+                || (other instanceof FilterPredicate // instanceof handles nulls
+                && keywords.equals(((FilterPredicate) other).keywords)); // state check
     }
 }
